Extract common components from login/sign up
