+++ frontend/html/login.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>VeritasAI - Login</title>
     <link rel="stylesheet" href="../styles/output.css" />
-    <script type="module" src="../scripts/login.js"></script>
+    <script type="module" src="../scripts/authentication/login.js"></script>
   </head>
   <body class="h-full">
     <div class="flex min-h-full flex-col justify-center space-y-10 px-6 py-12 lg:px-8">
